Added field validation to the Registration page
The fields on the registration page will validate on the blur event.
Some fields, like the username and email fields, will need to contact
the server in order to validate. As of this commit the validation has
been built into the frontend, but has yet to be completed on the server.