User Registration form w/ validation complete.
The validation on the registration form is fully functional. The server
accurately validates usernames and emails. The server still needs to be
configured to complete the registration process, including sending out
activation emails. An activation page must also be set up.

Lastly, a successful/failed registration attempt should be reported to
the user (i.e. an error message or success modal).