Activation works, Register process complete.
Beginning to add models and login state. The app checks the user's login
status when the Root component is mounted. It should be possible for a
user to register and activate an account. They can also have another
activation link sent to them if the original expires (after 72 hours).

There are error messages in case of internal server errors.