Added deploy script, virtual host config files.
I've moved the Apache mod_rewrite rules to the config directory, and
I've begun tracking the entire file, as opposed to tracking the rewrite
rules only. There are two virtual host config files: one for http, which
is responsible for redirecting the request to port 443 (SSL/HTTPS). That
virtual host file contains furthur rules for rewriting the URL so the
backend can understand the request.

The deploy script located in the config directory is to be used on the
production server after a git push. I think Git hooks might be useful to
automate the deployment.