Root component checks for user login before mount.
I originally had it run that method after mounting. The intended effect
of this change is to reduce the time it takes to load pages by allowing
the check to occur in the background as the components mount. Otherwise,
the components have to mount, then wait for the check, and then
re-render. This should allow the re-render to happen almost immediately
after the Root component mounts.