The Register page state resets after it mounts.
This fixes the issue where navigating away and then back to an
unsubmitted Register form will cause the user's entries to be lost, but
not the validation status of those old entries.

Alternatively, I could just as easily preserve the text as well. This
feature may be implemented in future.