Added the "Add Individual" page.
This is the page at `plantlo.gg/plants/individuals/add` that allows
users to add a single plant.

The submission of the form has not been handled yet, but the page itself
is accessible (a mod-rewrite rule has been added).

Styles and images for the page have been added as well, and the schema
has been updated to allow more NULL fields (to make adding a plant
easier).