Changed schema: table `varieties` renamed `taxa`.
I changed the name of the `varieties` table to `taxa` because the new
name reflects the nature of the table better than `varieties`.

A taxon is a taxonomic group of any rank. A variety, in botany, is a
specific rank below subspecies. Calling the table `varieties` implies
a list of actual varieties, but in this use case the table stores
full taxonomic names. Hence, the label `taxa.`