Taxon name no longer required to add an individual.
The idea is to make it as easy as possible to add an individual. The
only required field at this point is the individual's custom unique ID,
which allows the user to search for the plant using an identifier of
their choice. Of course, users can also assign simple nicknames to their
plants.