Added Plants, Individuals pages. Added list actions, filters.
The pages for displaying users' plants and individuals have been added,
both of which require lists. I've used a "List" component in these
pages, but it has not yet been implemented.

I have implemented constants which define lists' filters, sorts, and
layouts. I've also added the "ui" reducer to handle actions related to
the user-preferences of the various ui components.