Added PublicController::AddNewIndividual()
This is part of the server's public API. The client-side application
sends a JSON object to the server via this method so the record can be
persisted to the database.

The `src/lib` directory has also been added. It contains classes I've
written for the backend that don't fit nicely elsewhere.